Modify src/Adventure/Engine.hs
Added container objects.

In order to do this and distinguish containers from items I created a
more generic product type, `GameObject` which contains the common
fields between them and has a value of type `Object`.

The `Object` type is the sum which discriminates from `Item` game
objects and `containers`.

Because of the extra nesting I created smart constructors for `item`
and `container` types.

We may need to add lenses and prisms if it becomes a bit of a mess to
separate the two, but so far no need.