wip: Add ScrollTo WidgetRequest to Keystroke
Super experimental.  Trying to modify a copy of the Keystroke widget
from the Monomer library to send a `ScrollTo` message to our
scrollable scene list so that it can scroll to the _bottom_ when a new
item is added instead of the _top_ (as it does by default).

This change present hard-codes the `ScrollTo` message to a fixed
amount and manually targets the scroll area with a hard-coded `Path`.