Add $save command to save games
The game engine will ensure its directories exist at startup.

The save game command writes the save games to the path.

Changed the error message box to take text so that we could keep save
game errors out of the `GameError` type and still display
informational stuff to the user.