Add ItemPickedUp event to game event log
This change also includes a large refactor that:

 - breaks up the game engine module into submodules
 - adds a `MonadState GameState` to the context of `handle'`

The first part is required to break up circular imports when we made
"rewards" its own module.  This ended up trickling down and forcing
several other concepts to be packaged into modules to avoid circular
dependencies.

The second was required so that we didn't have to thread through the
game state explicitly down to handlers. This allowed a nice
implementation of an action `emitEvent` which more handlers can now
use to update the game event log.